Fix a few bugs and inconsistencies in grdblend
See #4943 for background.  There were two main problems:

When no -R was given but -I was given, we got a bogus message about -I being required and then it quit.  Now we use the given -I, if it is indeed missing then we bail.  Same situation with -r but not tested by @anbj.
The problem leading to most issues was that there was no check that the required arguemnt -Gfile was in fact given.  Things went rapidly downhill from there.
Closes #4943.