Improve grdblend checking of input arguments
See #4943 again.  Problem was that the user was giving a single file which we must assume is a blend file.  But it was not - it was a single grid.  OF course, it makes no sense to blend one grid, but trying to read it as a blendfile gave lots of irrelevant errors and warnings.
This PR checks for the special case of a single input file.  If it can be successfully determined to be a grid then we give an error message, otherwise we trust it is a blend file.  The final case not checked for is if no files are given and we read stdin.  However, having some user pipe a grid is very unlikely and derserves a bad ending.