Check result of gmt_get_ext before passing to other functions (#4981)
* Check result of gmt_get_ext before passing to other functions

See #4979.  The problem was we had the result of gmt_get_ext being passed directly to a strcmp function.  However, this file had no extension so gmt_get_ext return NULL and that crashed the strcmp.  There were two places in GMT where we did this.  Now replaced by a separate check if an extensino was found.  Closes #4979.

* Update gmt_esri_io.c