Address failure to output spatial coordinates when -a was used (#5005)
* Address failure to output spatial coordinates when -a was used

Using -a means output all aspatial items plus the spatial coordinates.  Unfortunately, GMT flagged the record as TEXT only and hence only wrote the aspatial trailing string...
I added a new test that checks if various combinations of -a arguments produce the correct results.  They do now, and if we mess up in the future then this test may trigger and warn us.

* Update check_a_opt.sh

* Update gmt_io.c