Force projection for images for global rectangular projections
See #4814 for background.  The problem is that when a rectangular projection is selected then there is no need to reproject a grid to plot it.  However, for images the reading does not shuffle the region, hence we do need to run it through the projection step.  Perhaps this may change in the future, but for now this catches those cases.  THis is why it worked when the OP specified the projection explicitly.  And it works for other projections like -Jh etc.