WIP Anim14
I share with you my idea for animation 14 to show the ability to make a movie with meca and coupe. There are many details that can be improve in the script but now I would like to know your general opinion about the figure before I go on improving the scripts and the animation.

About the data: I used the Global CMT web page and I add a column with the time.