Properly set min/max, lat-lilmit for global small circle
While this does not fix #4551 it fixes a minor bug that appears not to have caused any problems (yet). At least now, when checking if the four map corners are inside or outside the large solar polygon it returns the correct results.  We are just not handling the case correct yet.