Update admin/PS_orig_for_gs.* files (#5083)
* Update admin/PS_orig_for_gs.* files

Nothing to do with gs, but given our changes to themese and auto-font scaling, the plot we used has tiny changes and thus fails the gs-check.sh script.  New files created in this PR passes the test.  ALso, the updated PS uses the new current transparency codes instead of the obsolete stuff, so had to happen at some point anyway.

* Update ConfigDefault.cmake

* Update ConfigDefault.cmake