Ensure plot region needed for -Pf uses ISO formatting for absolute time (#5094)
See struggles in #5019.  The problem was I used the prevailing FORMAT_DATE_MAP and FORMAT_CLOCK_MAP settings to make the -R string but that is just wrong, of course.  For this purpose (making a -R argument to a module) we need FORMAT_DATE_OUT and FORMAT_CLOCK_OUT, as implemented in this PR.