Fix minor annoyances in the API (#5099)
Adding a few more checks to skip actions when input is a memory file, and also attempt to update the type field in the API object container when a matrix of known type is added.  None of these are expected to fix anything but is the right thing to do.