Incorrect number of output points in solar
When -M is used we wish to dump the coordinates of the polygon.  However,  the gmt_get_smallcircle function may add extra points if we cross Greenwhich or the Dateline, hence the input n points is the minimum returned.  Hence, the calling program must use S->n_rows and not the initial npts when writing the output.  No tests were affected though.