Use a separate grid index for area grid in grdfilter (#5107)
See this post for details.  The problem was that there was the assumtion that the input grid and the area-per-node grid could share the same grid index, but in this case the input grid was a matrix with no pad while the area grid is created internally and has the regualr pad.  hence different indices are needed.  Now works for me.