Failed to detect lack of modifier for -E
The grd2cpt -E option takes an optional +f<file> modifier, and for externals it is possible to just give +f and GMT_Encode_Options will fill in the memory reference.  However, in this case -E was given with no argument yet the API failed to detect this case and added a memory reference.  This PR is more careful to check that an option that may take modifiers did in fact do so before we consider this hookup.  WOrks for me but needs full Julia test.  I will run the mex 2017 paper tests.