Features.rst Issues in escape sequences examples
In the escape sequence section there is a list with some example. The last is "Acceleration (ms-2)." I add the backslash so the brackets are shown in the code (I try it) and complete the result with the "(ms@+-2@+)" (not sure if I edit it well).

Also, in the previous example (Manoa Straße) it is not clear what the escape sequences does in "Manoa". It was  Månoa?