Check if annotation phase implies a wider annotation format
When phase is specified it is possible that the phase-shifted annotation values may require more decimals than that we infer from the internal itself.  E.g., if interval is 0.5 and phase is 0.25 then annotations like 0.75, 1.25 etc will require 2 decimals, one more than inferred from the annotation interval.  This PR addresses this concern and closes #5152.