Ensure font/stroke color is reset internally (#5170)
GMT keeps track of the current colors for fill, font, and stroke so that we dont keep sending the same color request in PostScript.  However, when some activities have been placed under a gsave/grestore branch then onc we exit PostScript will revert to the previous setting, which is not remembered inside GMT.  The solution to this is to set the relevant colors to "not set" (-1, -1, -1) so that the very next call for a color will always be executed.  I added some comments in the new gmt_map_title function to remind us of this.
Closes #5156.