gpsgridder must set input columns before reading via -N (#5197)
The reading of the desired output points via -N takes place after we have read the 4 or 6 column data input.  However, we never reset the number of expected columns to be read when reading via -N, resulting in error messages that we dont have 4 (or 6) input columns, even though only 2 are needed.  This PR sets the expected number of columns to 2.