Prevent filename with space troubles in GMT_Call_Module
While GMT can read filenames with spaces if given in quotes, if these filenames are then passed as file arguments in strings given to GMT_Call_Module then those are lost and the called module will complain.  When GMT_Call_Module is used we replace spaces with ASCII 29 and do the opposite in the modules via GMT_File_Path.  Closes #5119.