Let grdsample adjust output increments before testing for region overlap (#5228)
* grdsample needs to finalize increments before determining region

If the increment is given in distance units then we must convert those to degrees before determining the suitable output grid region since that algorithm requires the grid spacing (which has not yet been converted from, say, meters to degrees.  This PR splits off the main calcualtions from gmt_RI_verify into a new function gmt_adjust_increments that can be called in situations, such as in grdsample, when we need to ensure this is set before going further.  Normally, this step is automatically taken care of in GMT_Create_Data but grdsample is an exception.

* Minor adjustments

* Set level to warning when creating a new grid

* Update ex16 to avoid prime dimensions