WIP Let grdtrack -C know if units are given as spherical arc lengths
THe problem was we always assumed a crossprofile length and along-track interval were in units like km or meters and thus always converted those to spherical degrees.  But if the input values already are in degrees we make chaos.  This PR fixes that as well as allowing an even number of nodes along a profile, meaning no value will be sampled at the main line.  Also added a check that will give a warning if length/spacing is more than 5% from an integer