Cartesian -L case failed as distances are squared but +d is not (#5253)
See forum for background.  The problem was that for simple Cartesian distances we save some processing time by using the squared distances instead of calling hypot.  But the +d distance was not squared, leading to huge mismatches.  We do the same for -C but here the squaring was taken care of.  Since gmt_near_lines are called in other modules the squaring is taken care of inside the gmtmap_near_a_line_cartesian function.  Also grdmath needed to call the right gmt_init_distaz so comparisions are correct.