Duplicate into reset pad instead of allocate twice
When an incoming grid has no pads and grdimage requires it we duplicate the incoming grid then expand it to add the pad.  However, this was done in a wasteful way where we first did an identicap duplicate, then allocated the final padded grid and copied over, thus having an extra grid in memory temporarily.  Turns out GMT_Duplicate_Data has a mode for this called GMT_DUPLICATE_RESET that does this in one step without adding the temporary memory.  Need @joa-quim and @seisman to see if this branch has any side-effects in GMT>jl and pyGMT since I cannot test that.