WIP Fix minor bugs in project -Z
There were a few minor issues in project -Z.

THe number of points created got computed in two different places and could differ due to roundoff, leading to an extra record with zeros.
The along-ellipse distances, which is approximate, could have an odd jump at the last record.

We can approve and merge this after 6.2.0 has been relesased.