grdblend could leak memory for pixel grids
If a tile is exactly outside the final grid region, then that check is slightly different for pixel and gridline-registered grids: For the same area the w/e/s/n are the same, but at the joint the pixel-grid has no nodes.  Hence pixel grids with the same north boundary as the south boundary in the final grid would be considered inside, memory allocated, but never touched and hence failed to free.