Wrong temp variable used when assigning tile weight (#5316)
Somewhere in the transcription from old GMT to the API a variable was renamed but not changed in one later line - probably by me.  So the tile weight for tile number two, if given via a blend file, would be set to zero.
This PR fixes that problem.  All tests pass meaning none of our tests explored this particular scenario, so I am adding the simple example I used to debug this.