Fix problems in gmt_parse|create_array
A failed if-test in gmt_parse_array meant that -Tdate1/date2 (in makecpt) would not be seen as temporal arguments and hence output would be in Cartesian seconds, not as a datetime stamp. Once fixed, it became clear that we also need to set the default time unit to the same as the system default [s] unless the user specified in increment with a specific unit.  This PR relates the the issue posted on the forum.