Let -f<col>s explicity set start of input trailing text (#5370)
* Let -f<col>s explicity set start of input trailing text

At least as long as GMT modules maintain record-by-record reading (i.e., as UNIX filters), we may not be able to reliable identify the start of trialing text if the first record has numbers in that column.  If that happens, then any trailing text in subsequent columns will be read as NaN and basically lost.  Using -f to explicitly state where trailing text starts makes the determination robust.  Closes #5368.

* Update gmt_init.c