Address shortcomings in grdcut -J
There was some confusion in grdcut.  For instance, it insisted that a region be given and that it be oblique when -J was set, but for many cases we just need -J and the full region of the grid.  If -J is given then we use -R if given and if not we use the grid region to set up the projection.  This works for the cases discussed on the forum, and all tests still pass.