Improve the hyphenation breaks in terminal
For things that are hyphenated in the text, e.g., half-circle, if that word needs to be split there should be no continuation symbols since it is a valid hyphenation.  This PR handles this better while looking out for option strings that should not break, i.e., we do not want -Qm to become - on line line and Qm on the other.