Turn off implicit output for externals if spectrum1d -T is given
spectrum1d has a long history, with initially writing individual files for each cross-spectra component.  Then we added the option to write a single table with all the items to stdout by default, but this could be turned off via -T.  Because of backwards compatibility we left the default action as it was, hence -T is needed to turn off a call to GMT_Write_Data.  Thus, GMT_ENcode_Options need to check for -T and if set it should deactivate the implicit output option.  I think it is very rare if any externals uses spectrum1d and does NOT want the result back as normal, but in case they do and use -T we do not want a nasty crash.