Fix regression for +sa auto grid-range scaling
When one wishes to save disk space by using an integer grid format, it is useful to have GMT automaticallly determine scales and offsets so that the maximum resolution of the data can be preserved despite being stored as integers.  However, we must have had a regression here since +sa lead to errors.  This PR fixes this problem.