Change name of remote file in sph2grd example
The sph2grd example references the remote file `@EGM96_to_360.txt`. There doesn't appear to be a file by that name on the GMT server, but there is "EGM96_to_36.txt". I'm assuming this should have been file referenced in the example. This pull requests changes the name of the remote file, and for consistency's sake changes the name of the outgrid as well.