Maintenance: use gmt_get|set_column_type
We have two functions that gets and sets the column type for input or output and given column, but we also have legacy code that manipulate the hidden array entries directly.  This PR just replaces some of the latter with the cleaner functions that makes the code easier to read.  To change in behavior and al ltests still pass.