Enable linear scaling of symbol proportional to moment
This PR implements Torsten Becker's suggestion of an option to scale symbol sizes proportional to moment and not just magnitude.  We do this by introducing some common constants in meca.h that sets the default earthquake parameters and then use those in coupe and meca.  Changing mode and setting parameters are done via new modifiers to -S.  We also deprecated -M to be a +m modifier in S instead.