Docs cleanup (#5623)
* Tiny fixes to the docs and usage.

For -q, explain that min/max has defaults if one is not given. Also, since we return an error if more tha one ~ is given then only one is allowed (not required). For -e, the +ffile is not a modifier but the way to select a file.  I do not like that since it is not consistent - will get back to that.  The -w option: Replace first colon with period.

* Update explain_-q_full.rst_

* Update doc/rst/source/explain_-q_full.rst_

Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>

* Update doc/rst/source/explain_-q_full.rst_

Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>

* Add limits to -q synopsis

Co-authored-by: Meghan Jones <meghanj@alum.mit.edu>