Enable gridlines when colormap -B implies it
See this forum post for background.  Turns out that when colormap used -B we did not draw gridlines since the Cartesian axis function used skipped gridlines as those were expected to be done later in the gmt_map_basemap funciton.  Except for colormap which does not use gmt_map_basemap and hence there was no call to plot gridlines.  This PR fixes that issue, updates a test that now failed (since it should have had gridlines but did not), and did some minor coding improvements to make it easier to understand the algorithms related to custom annotations.