Better estimation of background frame for colormap -Q
The calculation of the size of the background panel (-F) did not consider the differences in annotation widths when -Q (logarithmic) is used, as ponted out in this forum post.  This PR addresses this ommission by estimating the width of 10^value strings as well as the vertical increase in annotation height due to the superscripts.