Fix errors in sample1d for monotonically decreasing time
Turns out that while some parts of the new array machinery were aware of reversing time, others did not play ball, leading to errors.  This PR ensures that we check for this condition so that we do not pass max < min to functions that cannot handle that condition.
All tests pass.