Add linear moment scaling to coupe and meca (#5649)
* Enable linear scaling of symbol proportional to moment

This PR implements Torsten Becker's suggestion of an option to scale symbol sizes proportional to moment and not just magnitude.  We do this by introducing some common constants in meca.h that sets the default earthquake parameters and then use those in coupe and meca.  Changing mode and setting parameters are done via new modifiers to -S.  We also deprecated -M to be a +m modifier in S instead.

* Update scripts after first passing with -M

Now use -S .. +m modifier

* Eliminate symbol duplication

* Update CMakeLists.txt

* Improve usage

* Create seis_13.sh

* Finalize the test script for linear size

* polish