Improve events for trajectories
Trajectories are lines given by (x,y,t) and in an animation we wish to plot the portion of this line up to the current time.  This requires interpolation between the last data point older than current time and the next (future) point so that the end of the trajectory for the current time can be estimated.  This PR enables this feature and improves the documentation.