Two small issues in movie and events
In events, after we added the symbol size column, we forgot to set it to 1 by default (then to be modulated by the time curves).  So for plain animation with the default step function curves we ended up with a size of zero.
In movie, the resent updates to psconvert mean we need to insert a comma before adding N for color/outline of rrame since no longer a modifier of A.