Add v (vector) to custom symbol kit, plus handle azimuth vs angle (#5606)
* Add v (vector) to custom symbol kit, plus handle azimuth vs angle

It may be useful to draw a simple one-headed vactor as part of a custom symbol, but we had no such capability.  This is now added.  The vector has a head at the end and the size is derived from the current pen thickness.  Head outline is always drawn, and head may be filled if -G is used.  This PR also clarifies how to ensure angles are considered as Cartesian vs geographic azimuths, and allows a trailing a to an angular constant to mean azimuth.

* Add doc script and update docs

* Update gmt_plot.c

* Update figure

* Update gmt_plot.c