Consistently describe input grid name and modifiers (#5602)
* Show modifiers for ingrid in synopsis and usage

We tended to just say <ingrid> is the input grid, but there are several modifiers as well as a directive that can be used for reading and scaling data.  This is now documented in the synopsis and usage.

* Ensure all input grids are called <ingrid> consistently

* Ensure +d is recognized as grid modifier

* Update cookbook regarding +d

* Update gmt_synopsis.h

Add the GMT_OUTGRID constant as well. TO be used.

* Lay groundwork for outgrid syntax

* Do output separately

* Keep outgrid for separate PR

* More RST updates

use upper case ID for the gridfile id throughout.

* Include band and varname