Ensure we handle pixel grid with a dimension of 1 in row or col
The smallest grid possible is 1x1 if pixel-registration is selected.  However, we did not handle the assignment of increments etc when reading such grids and could crash for that rare case of 1x1 since we would read the vector (of length one) and then do the diff xy[1] - xy[0] to get the increment (which cannot be obtained that way if no vector).  We have the actual_range though and that is now used to get the increment.  This and a few other chekcs have been added to handle this case.  Also, I added language to the cookbook about the smallest grids possible in pixel (1x1) or gridline (2x2) format.