Update greenspline -C for internediate grids
Currently, the modifiers +m and +M are mutually exclusive and thus it is only possible to output intermediary grids for the cumulative or incremental solution.  However, since the main calculation time is in determining the solution, it is helpful to be able to write both types of grids at the same time.  This PR relaces the +m and +M with two new modifiers that both can be set: +c for cumulative grids and +i for incremental grids.  We now use the regular -Ggrdfile setting and automaticallly insert _inc_### or _cum_### before the file extension, and we start with eigenvalue 0 rather than calling it 1 (averything else in GMT starts at 0 so starting at 1 causes grief in batch and movie). The old +m|M modifier is honored in a backwards compatiple way with the original file template and mumbering scheme.