Reset PSL internals at end of building a completion function
The way subplot tags and movie labels and progress indicators work is that the PostScript needed to place these items are written as a "completion" PostSCript function and then the function is called at the end of the subplot or plot. One tricky thing is that if you for instance change a color in a completion function, the next PSL function requesting the same color will think it was already set, but the function has not yet run, hence we fail to select the color.  I had a partial sollution to this by resetting key internal parameters, but it was incomplete and failed in movie when several -L labels with the same color or pen outline were selected.  This PR gathers all the resetting statements in a new function (gmtplot_reset_PSL) that is now called at the end of building a completion funciton.  Tests still pass and my problem case also works.