Better recognition of C-style formats in movie (#5762)
The movie module had a lame check for format statements, such as checking for % and if there is a d or f etc.  But that mean "Altitude %4.0f km" would find a d but it is not part of the format and try to set an integer...
This PR adds a simple function that does the correct checks.