Centralize creation of templates and opening of temporary files
This PR simplifies the handling of temporary files.  Before, various parts of GMT would do this manually: Set up a template that contains the XXXXX pattern, call mktemp, then maybe append an extension. If a temp dir exists then taht was prepended.  Then we sometimes create the file, inc which case there is a choice between how this is done on Windows and Linux.  Now, al this is cerntralized into two new functions.