Get tighter w/e/s/n region for general perspective projection
For various reasons we need to determine the encompassing w/e/s/n rectangle in degrees for all proejction settings.  For the general perspective projection this is complicated by the fact that for some projection parameters the map window is the intersection of a rectangle and a sphere, yielding an odd rounded octagon.  We then search along this perimeter and determine the extreme longitude/latitude values. However, should any of the geographic poles be inside this polygon then we just set w = 0, e = 360 and update the max or min latitude.  That last step did not work for the genper because we ahd just copied/pasted what is done for other azimuthal projections.  This PR implements this correctly by determining of those pole points is inside the spherical polygon clip path.