Ensure we unlock if file is found (#5777)
This PR is an addendum to the recent file lock improvements: When the lock works and the process must wait, then when it is allowed to continue it will find the file already there.  However, in that case the (new) lock has been set and we should unlock before we return.