Adjust known answer following improved -JG search (#5790)
The result for this script was set before recent work on the JG rectangle search.  The old search always added 0.1 degrees to all borders (this was to ensure coverage in pscoast) but in other situations we do not want this.  Now that function has been generalized and takes an argument so that for typical lotting we do add 0.1 degree as before but in places like grdcut we do not.  This PR just updates the answers.