WIP Let grdcut -I operate on images
As a way to debug and implement subsets of images, I am letting grdcut -I get the ability to cut a subset out of an image.  This also makes it easier to debug this feature which is still not working in GMT (and affects grdimage, for instance, where the entire image must be read in).  Derives bits and pieces from stalled PR #4004.