Add useful warning to grdmix plus a test
I thought I was tracing down a bug in grdmix but it was just me forgetting to set -Ni.  Now, grdmix will warn us if you are trying to combine r,g,b grids and they exceed 1 (meaning they are likely in the 0-255 range). Added a test anyway to show that grdmix -D and -C return the original image.