Implement image subsets within GMT and via grdcut (#5798)
* WIP Let grdcut -I operate on images

As a way to debug and implement subsets of images, I am letting grdcut -I get the ability to cut a subset out of an image.  This also makes it easier to debug this feature which is still not working in GMT (and affects grdimage, for instance, where the entire image must be read in).  Derives bits and pieces from stalled PR #4004.

* Handle i->colormap as columns

* Update gmt_gdalread.c

* set no pad

* Finalize

* Remove -I and let grdcut figure out if it is an image instead

* Make sure we update "grid" type after reading image

* Implement west/west reading if straddling boundary

* Allow -S -F as well for images

* Update grdcut.c

* Update grdcut.c

* Better comments and add gmt_gdalread: to print statements

* Fix depadding in gmt_gdalwrite

* Update gmt_support.c

* Also unpack colors for pattern fills

* Want no subset of image here

* Update grdimage.c

* Finalize metadata explorer

* Improve BC handling for images

* BC stuff

* Use the data BC for images

* A few more things

* Update grdimage.c

* Update gmt_support.c

* Update gmt_gdalwrite.c

Minor initializations etc

* Update grdcut.c

Needed to turn on data padding in grdcut.

* Fix bad rounding

Better use irint to get nearest integer than rely on adding noise and cast to int.

* Update 3 PS files

* Update gmt_gdalread.c

* Better determination of tiff images

Add to gmt_raster_type since also needed in grdcut.

* Add extra test

Test based on Federico's Caribbean case plus a similar one straddling 180.  Passes.