Fix confusion between setting -r and what was set (#5820)
When surface -r was used we assumed this mean setting pixel registration.  However, when the syntax allowed -rg or -rp then we did not check the actual registration, only that -r had been set.  Of course, there is nothing to do if -rg (the default) is used, only if -rp.
Closes #5817.