Complete -D parsing update (#5827)
* Complete -D parsing update

When data cubes were introduced and we switched form +z to +d for setting data value (so +z can set uint for the z-axis in cubes), I allowed for backwards compatibility with +z for 2-D grids, but managed to not use the fix in the switch.  Today I ran into this bug - this PR fixes it.

* Update gmt_grdio.c

Also update the check of modifiers of course.