Prevent makecpt crash and set labels
When the number of category keys read from a file were fewer than entries in the CPT we got SEGV.  This PR instead exits with error in that case, and warns if we have too many keys that the rest will be ignored.  Also, when keys ar given from file we also use the keys as default labels.