Fix wrap bug and be quiet for some w/e/s/n rounding (#5861)
First, this PR fixes the crash reported in #5856.  Furthermore, it is very annoying to get the w/e/s/n rounding errors when the region originated from a polygon since you have no control over that.  This PR also adds more checking so that the strict GMT_CONV4_LIMIT slop checking is only reported as a warning when a w/e/s/n was given, not when -RFR is given, for example.