Avoid one-liner temporary symbol files
If an EPS is given as a custom symbol then we create a one-line custom symbol *.def file on the fly and add the single action of placing the EPS.  We then read that single-line custom file.  Two problems:
1.  Faster to just create internal symolb structure with that single action.
2. Clutter temp area with gmt_symbol?????.def files that are not removed.

This PR simply sets the custom symbol parameters in memory.  However, the gmtlib_convert_eps_to_def (now psxy_convert_eps_to_def) is still needed in psxy when we plot decorated lines since we must access the symbol via a separate call to plot.