Let psxy remove temporary symbol files used for decorating lines
For unclear reasons, we had a check that we only deleted the temporary GMT_symbol????.def file if it was not involving the decorated symbol file.  However, we plot the line first, then make a separate API call to psxy (from psxy) to plot the symbols.  Once that call returns the symbol file is no longer needed, so it should be deleted.

ALl tests still pass.