Add -B modifier to skip annotation(s) at end of axis
When composing multible frames or plots, expecially with y-annotations parallel to the axis, it is convenient to be able to skip plotting the annotations that land exactly at the end of an axis (or just one end).  This is difficult todo in GMT without changing -R (say -R0/0.9999 instead of -R0/1) but that is not a satisfactory solution.  The new axis modifier +e[l|h] will turn off both or just the lower or upper annotation if it lands at the end of the axis.  I have added a simple test script to illustrate the effect and ensure it works.