Enable sharing of SRTM tiles among multiple DEMs (#5870)
* Enable sharing of SRTM tiles among multiple DEMs

With the introduction of @earth_synbath there will be more than one data set (@earth_relief is the other) that for the highest resolutions will switch to SRTM 1 or 3 arc second tiles.  To avoid having to duplicate 41 Gb on the server as well as the user .gmt tree, I am adding code to extend the switch from earth_relief_01|3s to SRTMBL1|3 to also extend to other DEMs.  Currently that is only earth_synbath.  I do not anticipate more but if that should happen it will require a minor GMT verison update.  THis also means that only GMT 6.3 users can use @earth_synbath for regions where SRTM tiles may be involved.

* updates