Better srtm_tiles.nc allows for better checks (#5880)
The soon-to-be-updated srtm_tiles.nc now has three values: 0 means no SRTM data for this 1x1 tile, 1 if there is data but tile is partially in the ocean, and 2 if data and fully on land.  Previously, we only have values 0 and 1, with 1 indicating a file but we had no way of knowing if ocean filler would be needed.  Because of the need for backwards compatibility, the new file should work well with GMT 6.2 and earlier since the check was only on whether of not a node was zero or not.
This PR adds a few more checks, such as
Allowing for a future tile coverage grid to not be forced to be 0/360.
Add some enums to make the code easier to read