Use discovered w/e/s/n for reading subsets
When a remote grid or image is given without resolution we determine the correct resolution as well as possiblyl a subset region.  For tiled data we then only get the tiles in that region, but for the large images we failed to take advantage of knowing these subsets.  This PR makes it possible for the module to use this information.