Change libcurl connection timeout from default 300 s to 10s (#5891)
* Change connection timeout from default 300 s to 10s

Discovered yesterday when the Scripps CO2 site was down that GMT would wait 5 minutes before giving up on the connection.  This made me believe a test was failing and I wasted time tracking it down.  This PR changes the time-out (how long we wait to connect to a server before deciding it is not listening) to 10 seconds.

* Update gmt_remote.c