Pass GMT_GRID_NEEDS_PAD? flag to GMT_Read_Data if module requires a apd
Lots of GMT modules do not need a grid pad and are OK with reading a matrix from externals, but some requires a pad.  Those modules can pass the flags GMT_GRID_NEEDS_PAD1 or GMT_GRID_NEEDS_PAD2 and if the input data is by reference then we must switch to by duplication so a pad can be included.
Let us see if this helps with issue.