Improve min/max calculation in gmt_gdalread
This PR does three things:

It ensure we trust the GDAL min/max for floating point grids
It only updates z_min/z_max if ComputeRasterMinMax was called
It updates the commented out non-replacement code ta the end of the function to match the loops of the (now) piece-wise reading, but still commented out.

Tests pass.