Let movie and batch man pages warn of piping
Piping output form one GMT module into another may involve sub-shells which may not be behaving properly when we run multiple instances in movie and batch.  This PR adds a section that explains why it is not a good idea and that it can go wrong, and what a solution is.