Let GMT_GRAPHICS_DPU be the default in grdview -Qi (#5925)
* Let GMT_GRAPHICS_DPU be the default in grdview -Qi

The 100 dpi was set as a default decades ago when computers were slow and we were impatient.  Now that we have a GMT defaults that controls dpi we will use it. We let the tests stick with 100 where they previously just said -Qi so we dont have to update PS originals.

* Update viewpluscpt.sh

* Update ex32.ps