API: Switch to method from GMT_IS_REFERENCE to GMT_IS_DUPLICATE as needed (#5936)
* Switch to DUPLICATE if matrix is not of right shape and type

When a user matrix is used as a substitute for a grid and method is set to GMT_IS_REFERENCE, then there are some restrictions that apply, including no subset and that matrix is of the right type, is row-oriented, and has no complex components.  If that is not true then we switch method to GMT_IS_DUPLICATE and try again.  Same for user matrices passed as data cubes.

* Better comments and API fix

* Use a short function