Corect gmtspatial snafu and addtest
See this forum post for background.  Unfortunately, and clearly either while intoxicated or coffee-deprived, I pasted in the known geometry for the output into the slot for the data type, with the result that gmtspatial -D would exit pretty quickly due to this mistake.  Now added a test to ensure -D is tested at least once.