Fix wrap-around error in indexing for matrix-to-grid input
When desired subregion -Rw/e/s/n and a grid's internal region are off by 360 degrees relative to each other, we need to reconcile that before we compute column numbers since gridline vs pixel throws a wrench otherwise.  This PR does

Adjust teh internal grid header wesn to be compatible with the desired sub region
Also updates the objects method when we switch from reference to duplicate (I do'nt think this caused any issues but it is the right thing to).