Update the z min/max inside readonly subsets (#5948)
* Update the z min/max inside readonly subsets

When a matrix is passed by reference and there is no need to switch to duplication, we adjust the grid's pad to simulate a cut.  However, we never updated the header zmin/zmax to reflect the new data region.  This PR takes care of this for grids (and cubes).  Closes #5294.

* use GMT_LOCAL