Use GMT_NOTSET instead of -1 when that is what is meant (#5950)
We have lots of -1 in the code.  Sometimes it is a sign, sometimes is is a counter that will be incremented to 0 etc.  But often it is a return value that means not set.  We use GMT_NOTSET for this, but not everywhere.  TO make the code more readable I have replaced lots of -1 when I am sure they mean not set.

THings still works.