Let ex49 use mean for crustal ages (#5970)
* Let ex49 use mean for crustal ages

The data distribution of crustal ages within an age range is almost uniform, making a mode estimator a bad choice and very sensitive to noise (the Intel and arm hardware differ by over 1 Myr).  We fix this by asking for the mean crustal age instead which is unique.  This require an update PS so I also fixed some inconsistencies with the label and colorbar (now smaller, both have same shade box, etc.

* Upgrade to subplot

* argh

* Update ex49.rst

* Update ex49.ps

* Skip test of ex49