Avoid race condition when server table is refreshed
WHile we had some file locking for getting the cache hash table and the data server table, the latter is being brought in under a different name and then there is a renaming.  That section was not under file locking and the gmt_server_data.txt file can get clobbered by movie or batch.