Remember tile registration and increment (#5978)
* Remember tile registration and increment

We had a regression on this.  When I started to create the grid headers only based on internal information about a tiled data set, I forgot to update a few places in gmt_init.c so that the increment and registration was stored in the API structure.  This PR addresses this shortcoming.

* Add test

* Add test