Ensure secondary pens always set their transparency
See #5988 for background.  The issue is that the local psl_putcolor only checks if transparency is nonzero to issue a transparency setting, but for all we know another pen (as in the post above) has already declared transparency.  Hence, we should force the issuing of transparency even if there is none.