Fix remaining mem leak related to grdblend
We had an internal memory leak where the empty husk of the grid and header passed back from grdblend was dropped on the floor and left to be cleaned up when the process ends.  It is unseemly.  This PR allows GMT to free the memory as intended.

No tests affected but my memory leak messages have stopped (these requires special compiler flags to see).