Ensure correct domain/in is used for -G in blockm* (#6014)
We correctly handled increments given as -I10k in blockm* to determine the internal gridding setup, but then once that has been parsed the special flags used are reset. Hence, when the grid(s) needed for -G are created it only sees -I10 and we are off to SEGV.  This PR uses the initial correct header to created the next grids.