Set filler status before checking number of tiles (#6017)
This PR fixes an issue when @earth_relief_01s or 03s is requested but there are no 1s or 3s tiles inside the region and hence all we have is teh 15s filler tiles.  However, the decision to use filler came after we have already returned when we found no highres tiles.  The decision is now moved up a few lines.