Flip frame default for south pole aspect
See #6055 for background.  The issue was that the default of WeSn only makes sense in the northern hemisphere or when S is not the S pole.  This PR checks if we are doing a Sout polar azimuthal axpect with the S pole as the latitude parameter, in which csae we flip S to N in the axes defaults.  This also fixews the missing annotations in example 42 which also needed to be 1 cm taller.