Handle longer strings for netCDF history, remark and title
The length limits on these three were set back in the last century and because they are fixed array sizes in the GMT grid header they cannot be changed without breaking compatibility.  The solution is to add pointers to the hidden grid header structure that is not part of the API and hence store strings that exceed the limits here.  This PR does this juggling where we may need to use one of the other of the strings.