Rearranging code dropped a rewind? (#6088)
* Rearranging code dropped a rewind?

WHen the number of jobs are not known until we run the preflight script we must recreate the batch_init.sh file when taht is known.  This means we must read any -Iinclude script twice.  To do that we must rewind, not close, the include file pointer.  I do not recall having deleted that section but clearly it must have happened.  The script failed because we had already read the file and the file pointer was never rewound.
Now anim12.sh passes for me.

* Update batch.c