Add better checking in gmt_create_array
See #6093 for background.  The problem was that the bad date correctly was returned as NaN but then it was later used in min/max/inc to create an array and max was NaN.  This PR adds a check that neither min or max is NaN if that functions is called that way.  Closes #6093.